# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IConverterResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IConverterConfig": {
				"properties": {
					"source": {
						"type": "string"
					},
					"target": {
						"type": "string"
					}
				},
				"required": [
					"source",
					"target"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Boolean": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"IGitHubActionsConverterConfig": {
				"properties": {
					"source": {
						"type": "string"
					},
					"target": {
						"type": "string"
					},
					"singleFileTransformation": {
						"$ref": "#/components/schemas/Boolean"
					},
					"evaluationCreateYaml": {
						"$ref": "#/components/schemas/Boolean"
					}
				},
				"required": [
					"source",
					"target",
					"singleFileTransformation",
					"evaluationCreateYaml"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "master-thesis-generator",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Oliver Kabierschke"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/converter/jenkinstostalkcd": {
			"post": {
				"operationId": "ConvertJenkinsToStalkCd",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConverterResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Converter"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConverterConfig"
							}
						}
					}
				}
			}
		},
		"/converter/stalkcdtojenkins": {
			"post": {
				"operationId": "ConvertStalkCdToJenkins",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConverterResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Converter"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConverterConfig"
							}
						}
					}
				}
			}
		},
		"/converter/stalkcdtobpmn": {
			"post": {
				"operationId": "ConvertStalkCdToBPMN",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConverterResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Converter"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConverterConfig"
							}
						}
					}
				}
			}
		},
		"/converter/bpmntostalkcd": {
			"post": {
				"operationId": "ConvertBPMNToStalkCd",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConverterResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Converter"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConverterConfig"
							}
						}
					}
				}
			}
		},
		"/converter/bpmntojenkins": {
			"post": {
				"operationId": "ConvertBPMNToJenkins",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConverterResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Converter"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConverterConfig"
							}
						}
					}
				}
			}
		},
		"/converter/jenkinstogithubactions": {
			"post": {
				"operationId": "ConvertJenkinsToGitHubActions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConverterResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Converter"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGitHubActionsConverterConfig"
							}
						}
					}
				}
			}
		},
		"/converter/test": {
			"get": {
				"operationId": "Test",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConverterResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Converter"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}